backend/requirements.txt
django==4.1.7
asgiref==3.6.0
sqlparse==0.4.3
djangorestframework==3.14.0
pytz==2022.7.1
markdown==3.4.1
django-filter==22.1
django-cors-headers==3.14.0
drf-yasg==1.21.5